requirements.txt
asgiref==3.7.2
astroid==3.0.2
black==22.1.0
blue==0.9.1
click==8.1.7
colorama==0.4.6
dill==0.3.7
Django==5.0
flake8==4.0.1
isort==5.13.2
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.1.0
psutil==5.9.7
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
ruff==0.1.9
sqlparse==0.4.4
taskipy==1.12.2
tomli==2.0.1
tomlkit==0.12.3
